moved type guards
